requirements.txt
fastapi
uvicorn
openai
langchain
pandas
openpyxl
python-dotenv
requests
tqdm
sentence-transformers
psycopg2-binary